consolidate CategoryRollup.js under components/filters
